sqlite.lunar: two changes.
First change: If sqlite is requested and $MODULE_STATUS isn't an
sqlite database, migrate it to sqlite format automatically.

Second change: If an SQLite query fails, wait a short time and try again
(this is because occasionally multiple sqlites open up the database at
once).  This is of course an appalling hack.